Moved fpga specific functions from utils to targets/fpga_helper/fpga_utils
